Conditionally removing code handling "prev" pointer in the conscells.  Default if NOT to remove it yet, although with my quick test it seems to work... (And this REMOVES 'setf' routine from the scripts!)

+++ libsh/regex.h
@@ -35,7 +35,7 @@
    remains the value 0.  The bits are given in alphabetical order, and
    the definitions shifted by one from the previous bit; thus, when we
    add or remove a bit, only one other definition need change.  */
-typedef unsigned reg_syntax_t;
+typedef unsigned long reg_syntax_t;
 
 /* If this bit is not set, then \ inside a bracket expression is literal.
    If set, then such a \ quotes the following character.  */
